chore: update example create modal
