+++ examples/carbon-for-ibm-products/CreateModal/src/_index.scss
@@ -9,7 +9,6 @@ body {
 
 .app {
   display: flex;
-  min-height: calc(100vh - $spacing-09);
-  align-items: center;
   justify-content: center;
+  padding: $spacing-09;
 }
