--- examples/carbon-for-ibm-products/CreateModal/src/index.scss
@@ -1,53 +1,0 @@
-// Feature flags
-$feature-flags: (
-  enable-css-custom-properties: true,
-);
-
-// Carbon settings
-@import '@carbon/themes/scss/themes';
-@import '@carbon/layout/scss/layout';
-@import 'carbon-components/scss/components/dropdown/dropdown';
-@import 'carbon-components/scss/components/text-area/text-area';
-
-// Carbon for IBM Products package
-@import '@carbon/ibm-products/css/index.min.css';
-
-:root {
-  @include carbon--theme($carbon--theme, true);
-}
-
-body {
-  background-color: $ui-background;
-  color: $text-01;
-}
-
-$carbon--theme: $carbon--theme--white;
-
-.carbon-theme--white {
-  @include carbon--theme($carbon--theme--white, true);
-}
-
-.carbon-theme--g10 {
-  @include carbon--theme($carbon--theme--g10, true);
-}
-
-.carbon-theme--g90 {
-  @include carbon--theme($carbon--theme--g90, true);
-}
-
-.carbon-theme--g100 {
-  @include carbon--theme($carbon--theme--g100, true);
-}
-
-.app {
-  display: flex;
-  justify-content: center;
-  padding: $spacing-09;
-}
-
-.ccs-theme-dropdown {
-  position: fixed;
-  bottom: $spacing-05;
-  left: $spacing-05;
-  min-width: 11.5rem;
-}
